(feat) move position summary to data_types
# Conflicts:
#	hummingbot/strategy_v2/executors/executor_orchestrator.py
