+++ hummingbot/strategy_v2/executors/data_types.py
@@ -1,6 +1,7 @@
 import hashlib
 import random
 import time
+from decimal import Decimal
 
 import base58
 from pydantic.v1 import BaseModel, validator
@@ -34,3 +35,17 @@ class ConnectorPair(BaseModel):
         return self.connector_name in sorted(
             AllConnectorSettings.get_gateway_amm_connector_names()
         )
+
+
+class PositionSummary(BaseModel):
+    connector_name: str
+    trading_pair: str
+    volume_traded_quote: Decimal
+    amount: Decimal
+    breakeven_price: Decimal
+    unrealized_pnl_quote: Decimal
+    cum_fees_quote: Decimal
+
+    @property
+    def amount_quote(self) -> Decimal:
+        return self.amount * self.breakeven_price
